Fixed some problems with the TreeTableViewer that was causing changes to be left unsaved.  Also fixed a problem with trying to save the tree to DB after creating a new node and then deleting it all in one session.
git-svn-id: https://svn.code.sf.net/p/specify/code/trunk/Specify@1104 ce9bf4bf-376f-1848-a496-a1c9fe3b1a39

+++ src/edu/ku/brc/specify/treeutils/HibernateTreeDataServiceImpl.java
@@ -146,7 +146,10 @@ public class HibernateTreeDataServiceImpl <T extends Treeable<T,D,I>,
 			{
 				node.setParent(null);
 			}
-			session.delete(node);
+            if (node.getTreeId()!=null)
+            {
+                session.delete(node);
+            }
 		}
 		try
 		{
